refactor: move parseString.ts to args/
Issue: BTC-1351
